# swagger.yml
openapi: 3.0.0
info:
  title: Ephemeris
  version: 1.0.0
servers:
  - url: localhost:3000

tags:
  - name: Rendez-vous
paths:
  /connect:
    post:
      tags:
        - Rendez-vous
      summary: Se connecter à l'application
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: toto
                password:
                  type: string
                  example: totodu70
              required:
                - name
                - password
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    type: string
                    example: toto
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ
                  actions:
                      type: array
                      items:
                        $ref: '#/components/schemas/Actions'
      required: true
  /event/create:
    post:
      tags:
        - Rendez-vous
      summary: Créé un nouveau RendezVous
      requestBody:
        content: 
          application/json:
            schema:
              $ref: '#/components/schemas/RendezVous'
        required: true
      responses:
        '200':
          description: un objet RendezVous et les actions possibles
          content:
            application/json:
              schema: 
                type: object
                properties:
                  rdv:
                    $ref: '#/components/schemas/CompleteRendezVous'
                  actions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Actions'
        '400':
          description: format des données invalide
          content:
            application/json:
              schema: 
                type: object
                properties:
                  message:
                    type: string
                    example: Bad request
                  err:
                    type: object
                    propreties:
                      type: string
                      example: "..."
                      
  /event/get/{id}:
    get:
      tags:
        - Rendez-vous
      summary: Trouver un événement
      parameters:
        - name: id
          in: path
          description: ID de l'événement à trouver
          required: true
          schema:
            type: string
            example: '637740df4a124e6e61459d92'
      responses:
        '400':
          description: Format de l'id invalide
          content:
            application/json:
              schema: 
                type: object
                properties:
                  message:
                    type: string
                    example: Bad request
                  err:
                    type: object
                    propreties:
                      type: string
                      example: "..."
        '404':
          description: Aucun événement possédant cet id trouvé
          content:
            application/json:
              schema: 
                type: object
                properties:
                  message:
                    type: string
                    example: Ressource not found
                  err:
                    type: object
                    propreties:
                      type: string
                      example: "..."
        '200':
          description: Un événement a été trouvé
          content:
            application/json:
              schema: 
                type: object
                properties:
                  rdv:
                    $ref: '#/components/schemas/CompleteRendezVous'
                  actions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Actions'
                      
  /event/get-month/{year}/{month}:
    get:
      tags:
        - Rendez-vous
      summary: Trouver tous les événements pour un mois d'une année donnée
      parameters:
        - name: year
          in: path
          description: Année de l'événement à trouver
          required: true
          schema:
            type: integer
            example: 2022
        - name: month
          in: path
          description: Mois de l'événement à trouver
          required: true
          schema:
            type: integer
            example: 11
      responses:
        '400':
          description: Format de l'année ou du mois invalides
          content:
            application/json:
              schema: 
                type: object
                properties:
                  message:
                    type: string
                    example: Bad request
                  err:
                    type: object
                    propreties:
                      type: string
                      example: Date mal formée
        '200':
          description: des objet événement
          content:
            application/json:
              schema: 
                type: object
                properties:
                  rdvs:
                    type: array
                    items:
                      $ref: '#/components/schemas/CompleteRendezVous'
                  actions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Actions'
                
  /event/get-week/{year}/{week}:
    get:
      tags:
        - Rendez-vous
      summary: Trouver tous les événements pour une semaine d'une année donnée
      parameters:
        - name: year
          in: path
          description: Année de l'événement à trouver
          required: true
          schema:
            type: integer
            example: 2022
        - name: week
          in: path
          description: Semaine de l'événement à trouver
          required: true
          schema:
            type: integer
            example: 46
      responses:
        '400':
          description: Format de l'année ou du mois invalides
          content:
            application/json:
              schema: 
                type: object
                properties:
                  message:
                    type: string
                    example: Bad request
                  err:
                    type: object
                    propreties:
                      type: string
                      example: Date mal formée
        '200':
          description: des objets événements
          content:
            application/json:
              schema: 
                type: object
                properties:
                  rdvs:
                    type: array
                    items:
                      $ref: '#/components/schemas/CompleteRendezVous'
                  actions:
                    type: array
                    items:
                    $ref: '#/components/schemas/Actions'
                  
  /event/get-day/{year}/{month}/{day}:
    get:
      tags:
        - Rendez-vous
      summary: Trouver tous les événements pour un jour d'une semaine et d'une année donnée
      parameters:
        - name: year
          in: path
          description: Année de l'événement à trouver
          required: true
          schema:
            type: integer
            example: 2022
        - name: month
          in: path
          description: Mois de l'événement à trouver
          required: true
          schema:
            type: integer
            example: 11
        - name: day
          in: path
          description: Jour de l'événement à trouver
          required: true
          schema:
            type: integer
            example: 15
      responses:
        '400':
          description: Format du jour, du mois ou de l'année invalides
          content:
            application/json:
              schema: 
                type: object
                properties:
                  message:
                    type: string
                    example: Bad request
                  err:
                    type: object
                    propreties:
                      type: string
                      example: Date mal formée
        '200':
          description: des objets événements
          content:
            application/json:
              schema: 
                type: object
                properties:
                  rdvs:
                    type: array
                    items:
                      $ref: '#/components/schemas/CompleteRendezVous'
                  actions:
                    type: array
                    items:
                    $ref: '#/components/schemas/Actions'
                      
  /event/edit/{id}:
    post:
      tags:
        - Rendez-vous
      summary: Modifier un événement
      description: Le corps de la requête peut être composé de tous les champs présent dans un RendezVous
      parameters:
        - name: id
          in: path
          description: ID de l'événement à modifier
          required: true
          schema:
            type: integer
            example: '637740df4a124e6e61459d92'
      requestBody:
        content: 
          application/json:
            schema:
              type: object
              properties:
                title: 
                  type: string
                  example: "modification title"
                description:
                  type: string
                  example: "a better description"
      responses:
        '400':
          description: Format de l'id invalide
          content:
            application/json:
              schema: 
                type: object
                properties:
                  message:
                    type: string
                    example: Bad request
                  err:
                    type: object
                    propreties:
                      type: string
                      example: "..."
        '404':
          description: Aucun événement possédant cet id trouvé
          content:
            application/json:
              schema: 
                type: object
                properties:
                  message:
                    type: string
                    example: Ressource not found
                  err:
                    type: object
                    propreties:
                      type: string
                      example: "..."
        '401':
          description: Il faut être connecté pour utiliser cette fonctionnalitée
          content:
            application/json:
              schema: 
                type: object
                properties:
                  message:
                    type: string
                    example: Unauthorized
                  err:
                    type: object
                    propreties:
                      type: string
                      example: "..."
        '200':
          description: un objet RendezVous et les actions possibles
          content:
            application/json:
              schema: 
                type: object
                properties:
                  rdv:
                    $ref: '#/components/schemas/CompleteRendezVous'
                  actions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Actions'
  /event/delete/{id}:
    delete:
      tags:
        - Rendez-vous
      summary: Supprimer un événement
      parameters:
        - name: id
          in: path
          description: ID de l'événement à supprimer
          required: true
          schema:
            type: string
            example: '637740df4a124e6e61459d92'
      responses:
        '400':
          description: Format de l'id invalide
          content:
            application/json:
              schema: 
                type: object
                properties:
                  message:
                    type: string
                    example: Bad request
                  err:
                    type: object
                    propreties:
                      type: string
                      example: "..."
        '404':
          description: Aucun événement possédant cet id trouvé
          content:
            application/json:
              schema: 
                type: object
                properties:
                  message:
                    type: string
                    example: Ressource not found
                  err:
                    type: object
                    propreties:
                      type: string
                      example: "..."
        '401':
          description: Il faut être connecté pour utiliser cette fonctionnalitée
          content:
            application/json:
              schema: 
                type: object
                properties:
                  message:
                    type: string
                    example: Unauthorized
                  err:
                    type: object
                    propreties:
                      type: string
                      example: "..."
        '200':
          description: Un événement a été trouvé et supprimé
          content:
            application/json:
              schema: 
                type: object
                properties:
                  actions:
                    type: array
                    items:
                      type: object
                      properties:
                        link:
                          type: string
                          example: "/event/create/"
                        method: 
                          type: string
                          example: "post"

components:
  schemas:
    RendezVous:
      properties:
        title:
          type: string
          example: une choucroute chez mémé
        startDate:
          type: string
          example: "2022-10-25T11:30:00Z"
        endDate:
          type: string
          example: "2022-10-25T17:00:00Z"
        place:
          type: string
          example: "28 rue du chêne, 88000 Vosges"
        description:
          type: string
          example: "choucroute pour l'anniversaire de mémé (ne pas oublier le cadeau)"
      required:
        - title
        - date
    CompleteRendezVous:
      properties:
        title:
          type: string
          example: title
        startDate:
          type: string
          example: "2022-12-25T00:30:00Z"
        endDate:
          type: string
          example: "2022-12-26T01:00:00Z"
        place:
          type: string
          example: "there"
        description:
          type: string
          example: "this is a wonderful description"
        owner: 
          type: string
          example: "635149330eac27517446494c"
        _id: 
          type: string
          example: "63767163c1d5d7f8d523d2c1"
        __v:
          type: integer
          example: 0
    Actions:
      type: object
      properties:
        link:
          type: string
          example: "/.../..."
        method: 
          type: string
          example: "post"